move parsing library into `@lute/syntax`
