+++ definitions/luau.luau
@@ -1,17 +1,6 @@
 -- lute-lint-global-ignore(unused_variable)
-local cst = require("./syntax/cst")
 
 local luau = {}
-
---- Parses `source` as a Luau script and returns the resulting CST.
-function luau.parse(source: string): cst.CstParseResult
-	error("not implemented")
-end
-
---- Parses `source` as a single Luau expression and returns the resulting CST node.
-function luau.parseExpr(source: string): cst.CstExpr
-	error("not implemented")
-end
 
 --- Compiled Luau bytecode produced by `luau.compile`.
 export type Bytecode = { bytecode: string }
