+++ lute/std/libs/syntax/parser.luau
@@ -2,24 +2,24 @@
 -- @std/syntax/parser
 -- Parser for Luau source code into Luau CST nodes
 
-local luau = require("@lute/luau")
+local luteParser = require("@lute/syntax/parser")
 local types = require("./types")
 
 local parser = {}
 
 --- Parses Luau source code into a CstStatBlock
 function parser.parseBlock(source: string): types.CstStatBlock
-	return luau.parse(source).root
+	return luteParser.parse(source).root
 end
 
 --- Parses `source` as a single Luau expression and returns the resulting `CstExpr`.
 function parser.parseExpr(source: string): types.CstExpr
-	return luau.parseExpr(source)
+	return luteParser.parseExpr(source)
 end
 
 --- Parses `source` as a complete Luau file and returns the full `CstParseResult`, including the root block and EOF token.
 function parser.parse(source: string): types.CstParseResult
-	return luau.parse(source)
+	return luteParser.parse(source)
 end
 
 return table.freeze(parser)
